+++ safe/builtins/src/mkdirat/parse/mod.rs
@@ -41,7 +41,7 @@ pub fn mkdirat_parse<'a>(args: &[&'a CStr]) -> Result<MkdiratConfig<'a>, Report<
                     .change_context(BuiltinError::InvalidArgument("mode"))
                     .attach_opaque(Suggestion(
                         "Use octal without prefix (e.g. 755) or hex with 0x prefix (e.g. 0x1ff)",
-                    ))? as u32;
+                    ))?;
             }
             b"--resolve" => {
                 let s = crate::argparse::next_val(args, &mut i, val)?;
